Moved graphics to /tmp temporarily, and removed graphics debugging comments.
