+++ game.h
@@ -5,13 +5,13 @@
 #include "board.h"
 #include "player.h"
 #include "human.h"
-#include "graphDisplay.h"
+// #include "graphDisplay.h"
 
 // Wrapper class for game logic
 class Game {
     public:
         std::unique_ptr<TextDisplay> td = std::make_unique<TextDisplay> ();
-        std::unique_ptr<GraphDisplay> gd = std::make_unique<GraphDisplay>();
+        // std::unique_ptr<GraphDisplay> gd = std::make_unique<GraphDisplay>();
         std::unique_ptr<Board> b = std::make_unique<Board>();
         std::shared_ptr<Player> p1;
         std::shared_ptr<Player> p2;
