+++ main.cc
@@ -3,11 +3,11 @@
 #include "game.h"
 #include "enums.h"
 using namespace std;
- 
+
 
 int main(int argc, char *argv[]) {
 	// TODO: PARSE FLAGS
- 
+
 	// TODO: PARSE SETUP
     cout << "WELCOME TO CHESS! ";
     cout << "PLEASE ENTER ONE OF THE FOLLOWING COMMANDS TO BEGIN:" << endl;
@@ -29,11 +29,11 @@ int main(int argc, char *argv[]) {
         else if (command == "game") {
             if (!g->customSetup) {
                 g->td->setupTiles(g->isWhiteTurn, false);
-                g->gd->setupTiles(g->isWhiteTurn, false);
+                // g->gd->setupTiles(g->isWhiteTurn, false);
                 g->b->init(g->td->tiles); // default setup is always valid
             }
             g->b->attach(g->td.get());
-            g->b->attach(g->gd.get());
+            // g->b->attach(g->gd.get());
             g->runGame();
             cout << "Final Score:" << endl;
             cout << "White: " << g->p1->score << endl;
@@ -41,10 +41,9 @@ int main(int argc, char *argv[]) {
         }
         else if (command == "setup") {
             cout << "STARTING SETUP" << endl << endl;
-            if (g->td->setupTiles(g->isWhiteTurn, true) &&    
+            if (g->td->setupTiles(g->isWhiteTurn, true) &&
                 g->b->init(g->td->tiles)) {
-                // g->gd->setupTiles(g->td->tiles);
-                g->gd->updateTiles(g->td->tiles);
+                // g->gd->updateTiles(g->td->tiles);
                 g->customSetup = true;
             } else {
                 cout << "Invalid setup." << endl << endl;
