--- tmp/window.d
@@ -1,1 +1,0 @@
-window.o: window.cc window.h
