+++ tmp/window.h
@@ -23,8 +23,10 @@ public:
 	void fillRectangle(int x, int y, int width, int height, int colour=Black);
 
 	// Draws a string
-	void drawString(int x, int y, std::string msg);
+	// void drawString(int x, int y, std::string msg);
+	void drawString(int x, int y, std::string msg,int colour = Black);
 
+	void printMessage(int x, int y, const std::string& msg, int colour, XFontStruct& f);
 };
 
 #endif
