--- window.h
@@ -1,32 +1,0 @@
-#ifndef __WINDOW_H__
-#define __WINDOW_H__
-#include <X11/Xlib.h>
-#include <iostream>
-#include <string>
-
-class Xwindow {
-	Display *d;
-	Window w;
-	int s;
-	GC gc;
-	unsigned long colours[10];
-
-public:
-	Xwindow(int width=500, int height=500);  // Constructor; displays the window.
-	~Xwindow();                              // Destructor; destroys the window.
-	Xwindow(const Xwindow&) = delete;
-	Xwindow &operator=(const Xwindow&) = delete;
-
-	enum {White=0, Black, Red, Green, Blue}; // Available colours.
-
-	// Draws a rectangle
-	void fillRectangle(int x, int y, int width, int height, int colour=Black);
-
-	// Draws a string
-	// void drawString(int x, int y, std::string msg);
-	void drawString(int x, int y, std::string msg,int colour = Black);
-
-	void printMessage(int x, int y, const std::string& msg, int colour, XFontStruct& f);
-};
-
-#endif
